Use size_t for capacity instead of int

+++ src/chan.h
@@ -33,9 +33,8 @@ typedef struct chan_t
 
 // Allocates and returns a new channel. The capacity specifies whether the
 // channel should be buffered or not. A capacity of 0 will create an unbuffered
-// channel. Sets errno and returns NULL if initialization failed or the
-// capacity is less than 0.
-chan_t* chan_init(int capacity);
+// channel. Sets errno and returns NULL if initialization failed.
+chan_t* chan_init(size_t capacity);
 
 // Releases the channel resources.
 void chan_dispose(chan_t* chan);
